+++ CHANGELOG.md
@@ -2,12 +2,24 @@
 
 All Notable changes to `drupal/opening-hours` module.
 
+## [8.x-1.0-aplha2]
+
+### Added
+
+* DMOH-22: Added the opening hours widget formatter.
+
+### Fixed
+
+* Fixed package name.
+
 ## [8.x-1.0-aplha1]
 
 ### Added
+
 * DMOH-19: Added service configuration form.
 * DMOH-20: Added the opening hours field type.
 * DMOH-21: Added the opening hours field widget.
 
+[8.x-1.0-aplha2]: https://github.com/StadGent/drupal_module_opening-hours/compare/8.x-1.0-aplha1...8.x-1.0-aplha2
 [8.x-1.0-aplha1]: https://github.com/StadGent/drupal_module_opening-hours/releases/tag/8.x-1.0-aplha1
 [Unreleased]: https://github.com/StadGent/drupal_module_opening-hours/compare/master...develop